making some adjustments
