+++ pizza_flutter/lib/app/app.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+import '/app/ui/intro_screen.dart';
+
+class PizzaApp extends StatelessWidget {
+  const PizzaApp({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      title: 'Pizza',
+      themeMode: ThemeMode.system,
+      theme: ThemeData.light(),
+      darkTheme: ThemeData.dark(),
+      home: const IntroScreen(),
+    );
+  }
+}
