+++ pizza_flutter/lib/app/ui/home.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+
+class Home extends StatefulWidget {
+  const Home({Key? key}) : super(key: key);
+
+  @override
+  State<Home> createState() => _HomeState();
+}
+
+class _HomeState extends State<Home> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Column(
+        children: <Widget>[
+          Row(
+            children: const <Widget>[
+              Text('Nevada, US')
+            ],
+          )
+        ],
+      ),
+    );
+  }
+}
