+++ pizza_flutter/lib/main.dart
@@ -1,20 +1,6 @@
 import 'package:flutter/material.dart';
-import '../ui/intro_screen.dart';
+import 'app/app.dart';
 
 void main() => runApp(const PizzaApp());
 
-class PizzaApp extends StatelessWidget {
-  const PizzaApp({Key? key}) : super(key: key);
 
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      title: 'Pizza',
-      themeMode: ThemeMode.system,
-      theme: ThemeData.light(),
-      darkTheme: ThemeData.dark(),
-      home: const IntroScreen(),
-    );
-  }
-}
